🆕 chore(.editorconfig): add .editorconfig file The .editorconfig file has been added to the project root directory. It defines the coding style and formatting rules for the project. The rules include using spaces for indentation, UTF-8 character encoding, trimming trailing whitespace, and inserting a final newline. The max_line_length rule is set to off to avoid issues with long strings or URLs. The CHANGELOG.md file has a specific rule to ignore the indent_size rule.
🔍 chore(.gitignore): add gitignore file generated by toptal.com
This commit adds a .gitignore file generated by toptal.com. The file contains a list of files and directories that should be ignored by git. This helps to keep the repository clean and avoid committing unnecessary files.

📝 docs(query.mongodb): add query to find count of zips in NY state
This commit adds a new query to the `query.mongodb` file that finds the count of zip codes in the state of New York. The `use` command is used to switch to the `sample_training` database before running the query.